pyproject.toml
[tool.poetry]
name = "smarty-bot"
version = "0.1.0"
description = ""
authors = ["Andrei <tech@anlab.io>"]
readme = "README.md"
packages = [{include = "src/telegram_bot"}, {include = "src/ton_analyze"}]


[tool.poetry.dependencies]
python = "^3.10"
pytest = "^7.3.1"
telethon = "^1.36.0"
pytonapi = "0.3.8"
python-dotenv = "^1.0.1"
sqlmodel = "^0.0.22"
alembic = "^1.13.2"
rich = "^13.8.0"
psycopg2-binary = "^2.9.10"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
